moved code to RaceTab.tsx and deleted LapTab.tsx
